+++ cpp/box/box.h
@@ -1,13 +1,13 @@
 // Copyright (c) 2010-2018 The Regents of the University of Michigan
 // This file is part of the freud project, released under the BSD 3-Clause License.
 
-#include <iostream>
-#include <sstream>
 #include <cassert>
-#include <stdexcept>
-#include "HOOMDMath.h"
-#include "VectorMath.h"
+#include <iostream>
 #include <math.h>
+#include <sstream>
+#include <stdexcept>
+
+#include "VectorMath.h"
 
 #ifndef _BOX_H__
 #define _BOX_H__
@@ -40,7 +40,7 @@ class Box
             {
             m_2d = false; // Assign before calling setL!
             setL(0,0,0);
-            m_periodic = make_uchar3(1,1,1);
+            m_periodic = vec3<bool>(true, true, true);
             m_xy = m_xz = m_yz = 0;
             }
 
@@ -49,7 +49,7 @@ class Box
             {
             m_2d = _2d; //Assign before calling setL!
             setL(L,L,L);
-            m_periodic = make_uchar3(1,1,1);
+            m_periodic = vec3<bool>(true, true, true);
             m_xy = m_xz = m_yz = 0;
             }
 
@@ -58,7 +58,7 @@ class Box
             {
             m_2d = _2d;  // Assign before calling setL!
             setL(Lx,Ly,Lz);
-            m_periodic = make_uchar3(1,1,1);
+            m_periodic = vec3<bool>(true, true, true);
             m_xy = m_xz = m_yz = 0;
             }
 
@@ -67,7 +67,7 @@ class Box
             {
             m_2d = _2d;  // Assign before calling setL!
             setL(Lx,Ly,Lz);
-            m_periodic = make_uchar3(1,1,1);
+            m_periodic = vec3<bool>(true, true, true);
             m_xy = xy; m_xz = xz; m_yz = yz;
             }
 
@@ -322,28 +322,13 @@ class Box
                 }
            }
 
-        //! Wrap a vector back into the box. Legacy float3 version. Deprecated.
-        /*! \param w Vector to wrap, updated to the minimum image obeying the periodic settings
-         *  \param img Image of the vector, updated to reflect the new image
-         *  \param flags Vector of flags to force wrapping along certain directions
-         *  \post \a img and \a v are updated appropriately
-         *  \note \a v must not extend more than 1 image beyond the box
-         */
-        void wrap(float3& w, int3& img, char3 flags = make_char3(0,0,0)) const
-            {
-                vec3<float> tempcopy;
-                tempcopy.x = w.x; tempcopy.y = w.y; tempcopy.z = w.z;
-                wrap(tempcopy, img, flags);
-                w.x = tempcopy.x; w.y = tempcopy.y; w.z=tempcopy.z;
-            }
-
         void minimalwrap(vec3<float>& w) const
             {
             vec3<float> L = getL();
 
             if (m_periodic.x)
                 {
-                float tilt_x = (m_xz - m_xy*m_yz) * w.z + m_xy * w.y;
+                float tilt_x = (m_xz - m_xy * m_yz) * w.z + m_xy * w.y;
                 if (w.x >= m_hi.x + tilt_x)
                     {
                     w.x -= L.x;
@@ -400,17 +385,6 @@ class Box
             vec3<float> tempcopy = w;
             minimalwrap(tempcopy);
             return tempcopy;
-            }
-
-        float3 wrap(const float3& w, const char3 flags = make_char3(0,0,0)) const
-            {
-               vec3<float> tempcopy;
-               tempcopy.x = w.x; tempcopy.y = w.y; tempcopy.z = w.z;
-               int3 img = getImage(tempcopy);
-               wrap(tempcopy, img, flags);
-               float3 wrapped;
-               wrapped.x = tempcopy.x; wrapped.y = tempcopy.y; wrapped.z=tempcopy.z;
-               return wrapped;
             }
 
         //! Unwrap a given position to its "real" location
@@ -488,32 +462,32 @@ class Box
          *  \post Period flags are set to \a periodic
          *  \note It is invalid to set 1 for a periodic dimension where lo != -hi. This error is not checked for.
          */
-        void setPeriodic(uchar3 periodic)
+        void setPeriodic(vec3<bool> periodic)
             {
             m_periodic = periodic;
             }
 
         void setPeriodic(bool x, bool y, bool z)
             {
-            m_periodic = make_uchar3(x, y, z);
+            m_periodic = vec3<bool>(x, y, z);
             }
 
         //! Set the periodic flag along x
         void setPeriodicX(bool x)
             {
-            m_periodic = make_uchar3(x, m_periodic.y, m_periodic.z);
+            m_periodic = vec3<bool>(x, m_periodic.y, m_periodic.z);
             }
 
         //! Set the periodic flag along y
         void setPeriodicY(bool y)
             {
-            m_periodic = make_uchar3(m_periodic.x, y, m_periodic.z);
+            m_periodic = vec3<bool>(m_periodic.x, y, m_periodic.z);
             }
 
         //! Set the periodic flag along z
         void setPeriodicZ(bool z)
             {
-            m_periodic = make_uchar3(m_periodic.x, m_periodic.y, z);
+            m_periodic = vec3<bool>(m_periodic.x, m_periodic.y, z);
             }
 
         //! Get the periodic flags
@@ -548,7 +522,7 @@ class Box
         float m_xy;            //!< xy tilt factor
         float m_xz;            //!< xz tilt factor
         float m_yz;            //!< yz tilt factor
-        uchar3 m_periodic;     //!< 0/1 to determine if the box is periodic in each direction
+        vec3<bool> m_periodic; //!< 0/1 to determine if the box is periodic in each direction
         bool m_2d;             //!< Specify whether box is 2D.
     };