+++ cpp/order/Pairing2D.cc
@@ -1,15 +1,14 @@
 // Copyright (c) 2010-2018 The Regents of the University of Michigan
 // This file is part of the freud project, released under the BSD 3-Clause License.
 
-#include "Pairing2D.h"
-
+#include <complex>
 #include <stdexcept>
+#include <tbb/tbb.h>
 #ifdef __SSE2__
 #include <emmintrin.h>
 #endif
 
-#include <complex>
-#include <tbb/tbb.h>
+#include "Pairing2D.h"
 
 using namespace std;
 using namespace tbb;