+++ cpp/pmft/PMFTXY2D.cc
@@ -1,13 +1,12 @@
 // Copyright (c) 2010-2018 The Regents of the University of Michigan
 // This file is part of the freud project, released under the BSD 3-Clause License.
 
-#include "PMFTXY2D.h"
-#include "ScopedGILRelease.h"
-
 #include <stdexcept>
 #ifdef __SSE2__
 #include <emmintrin.h>
 #endif
+
+#include "PMFTXY2D.h"
 
 using namespace std;
 using namespace tbb;