+++ cpp/shapesplit/shapesplit.h
@@ -1,18 +1,12 @@
 // Copyright (c) 2010-2018 The Regents of the University of Michigan
 // This file is part of the freud project, released under the BSD 3-Clause License.
 
-#include <tbb/tbb.h>
-#include <ostream>
-
-// work around nasty issue where python #defines isalpha, toupper, etc....
-#undef __APPLE__
-#include <Python.h>
-#define __APPLE__
-
 #include <memory>
+#include <ostream>
+#include <tbb/tbb.h>
 
-#include "VectorMath.h"
 #include "box.h"
+#include "VectorMath.h"
 #include "Index1D.h"
 
 #ifndef _SHAPESPLIT_H__