+++ freud/density.py
@@ -13,7 +13,3 @@ from ._freud import LocalDensity
 from ._freud import RDF
 from ._freud import ComplexCF
 from ._freud import FloatCF
-
-# continue handling deprecated API
-from ._freud import ComplexCF as ComplexWRDF
-from ._freud import FloatCF as FloatWRDF